+++ metrics/__init__.py
@@ -3,15 +3,10 @@ import logging
 import os
 import sys
 import time
-import database
+from metrics.config import database
 
-import cpu_metrics as cpu
-import memory_metrics as memory
-import disk_metrics as disk
-import network_metrics as network
-import sensors_metrics as sensors
-import others_metrics as others
-
+from metrics.scripts_metrics import sensors_metrics as sensors, memory_metrics as memory, cpu_metrics as cpu, \
+    disk_metrics as disk, network_metrics as network, others_metrics as others
 
 logging.basicConfig(level=logging.INFO, filename=f"{os.getcwd()}/metrics/metrics.log", filemode='w',
                     format='%(message)s')
