+++ metrics/scripts_metrics/sensors_metrics.py
